requirements.txt
python-dotenv==1.1.1
streamlit==1.47.1
langchain-community==0.3.27
langchain-text-splitters==0.3.9
langchain-groq==0.3.6
unstructured==0.18.11
unstructured[pdf]==0.18.11
faiss-cpu==1.11.0
transformers==4.54.0
sentence-transformers==5.0.0
nltk==3.9.1